requirements.txt
# Core ML libraries
transformers>=4.56.2
datasets>=4.1.1
accelerate>=1.10.1
peft>=0.17.1
safetensors>=0.6.2
bitsandbytes>=0.42.0
torch>=2.8.0

# Document retrieval
sentence-transformers>=5.1.1
faiss-cpu>=1.12.0
scikit-learn>=1.7.2

# Data processing
pdfplumber>=0.11.7
numpy>=2.3.3
pandas>=2.3.2

# Utilities
tqdm>=4.67.1
requests>=2.32.5
